@matthewturk 2023-11-01
Nice!  I think it might be interesting at a future time to explore the impact of using a memory map, to see if that can farm out the skipping/paging/seeking to the OS.