@matthewturk 2023-11-02
Perhaps this is a silly question, but have we evaluated the speed costs of computing the (absolute) offsets ahead-of-time for all fields, then doing absolute seeks sorted by the position in the file?  Rather than doing the cumulative seek-by-field-size?