@matthewturk 2023-11-03
I know we already sort by that -- I just meant to eliminate *some* of the overhead in computing the field skips etc by precomputing the exact position offsets and seeking directly to them.