@cphyc 2023-11-03
Oh, I see. Let me illustrate to know if I understood your point correctly:
```
- iterate over levels
  - iterate over CPUs
	  1.   do an absolute seek
	  2.   do a relative seek to the first field to be read
	  3.   eventually read the ones immediately after
	  4.   do a relative seek to the next batch of fields to be read
	  5.   eventually read the ones immediately after
	  6.   [...]
 	  n.   do a relative seek to the last batch of fields to be read
	  n+1. eventually, read the ones immediately after
```

Following your comment, I think I optimized the hell out of the reader by putting together the first absolute seek + first relative seek (1 + 2).
With 7a953bcbe, I'm now limiting the number of jumps to its minimum. We _could_ compute ahead of time the value of the `skip_len`, but essentially each entry will have a unique value, so we would just be displacing the computation from being inline to being out of the loop (with small extra memory footprint). Does this make sense?