@matthewturk 2023-11-03
Not quite, and I also don't want to side-track you too much.  My idea was simply to compute the absolute of each field, but I am now realizing that one big speedup we'd miss would be reading stride 1 data of size greater than a single field, whereas with my question it would require reading a single field at once (even if the seeks were not minimized).  So forget it -- it's pointless!